Updated site.htm
Added/changed lines to mimic the most updated version that allows blocking text when entering passwords as well as the form method being post so information is not displayed. I have excluded the JavaScript portion that forces a user to enter both a username and password. This can be added at a later time.